AUDIO: MobC00194796
[Defect / Enhancement / New Feature]
MobC00194796

[Problem]
RheaBerri ICS: Distortion observed in Audio at Max Volume

[Solution]
Besides the parm_audio.txt change on CP side, (the CL that changed mixer output find gain),
also need this AP code change to correctly read mixer output fine gain from SYSPARM.

[Reviewers]
audio team

[Testing]
    [Test Cases]
play music and read audio SYSPARM, exam mixer gain in HW.

    [Label/CL# Synced To]
3-15 workspace.

    [Platform]
android_rheastone_edn00

[BROADCOM INTERNAL NOTES]

+++ sound/caph/audio_controller/audio_controller.c
@@ -3525,9 +3525,9 @@ static void fillUserVolSetting(AudioMode_t mode, AudioApp_t app)
 
 	p = &(AudParmP()[mode + app * AUDIO_MODE_NUMBER]);
 
-	user_vol_setting[app][mode].L = p->srcmixer_output_fine_gain_l;
+	user_vol_setting[app][mode].L = (short)p->srcmixer_output_fine_gain_l;
 	user_vol_setting[app][mode].L *= 25;
-	user_vol_setting[app][mode].R = p->srcmixer_output_fine_gain_r;
+	user_vol_setting[app][mode].R = (short)p->srcmixer_output_fine_gain_r;
 	user_vol_setting[app][mode].R *= 25;
 	user_vol_setting[app][mode].valid = TRUE;
 }
